fix: add guard logic in auction, bid page
