+++ src/pages/AuctionPage/AuctionPage.jsx
@@ -6,7 +6,8 @@ import { RxDotFilled } from 'react-icons/rx';
 import { useQuery } from '@tanstack/react-query';
 // import useModalStore from '../../stores/modalStore';
 import useLoginStore from '../../stores/loginStore';
-import PaymentConfirmModal from '../../components/PaymentConfirmModal';
+import useModalStore from '../../stores/modalStore';
+import PaymentConfirmModal from './PaymentConfirmModal';
 import PaymentModal from '../../components/PaymentModal';
 import {
   isWithinFiveMinute,
@@ -22,7 +23,7 @@ export default function AuctionPage() {
   const { auctionId } = useParams();
   const navigate = useNavigate();
   const { loggedIn, userId } = useLoginStore();
-  // const { openModal } = useModalStore();
+  const { openModal } = useModalStore();
 
   const [showDeleteConfirmModal, setShowDeleteConfirmModal] = useState(false);
   const [showBidHistoryModal, setShowBidHistoryModal] = useState(false);
@@ -123,7 +124,12 @@ export default function AuctionPage() {
                       type="button"
                       className="inline-block rounded bg-danger px-6 pb-2 pt-2.5 text-xs font-medium uppercase leading-normal text-white shadow-[0_4px_9px_-4px_#3b71ca] transition duration-150 ease-in-out hover:bg-danger-600 hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:bg-primary-600 focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:outline-none focus:ring-0 active:bg-primary-700 active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] dark:shadow-[0_4px_9px_-4px_rgba(59,113,202,0.5)] dark:hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)]"
                       onClick={() => {
-                        setShowConfirmModal(true);
+                        if (loggedIn) {
+                          setShowConfirmModal(true);
+                        } else {
+                          alert('로그인이 필요합니다.');
+                          openModal();
+                        }
                       }}
                     >
                       즉시결제
